+++ Project Files/Source/wdsp/utilities.c
@@ -264,7 +264,7 @@ void __cdecl CalccPrintSamples (void *pargs)
 {
 	int i;
 	double env_tx, env_rx;
-	int channel = (int)pargs;
+	int channel = (int)(uintptr_t)pargs;
 	CALCC a = txa[channel].calcc.p;
 	FILE* file = fopen("samples.txt", "w");
 	fprintf (file, "\n");
@@ -283,7 +283,7 @@ void __cdecl CalccPrintSamples (void *pargs)
 
 void doCalccPrintSamples(int channel)
 {	// no sample buffering - use in single cal mode
-	_beginthread(CalccPrintSamples, 0, (void *)channel);
+	_beginthread(CalccPrintSamples, 0, (void *)(uintptr_t)channel);
 }
 
 void print_anb_parms (const char* filename, ANB a)
